Play around with dumping arrays of bytes.
This isn't .. well, the cleanest/best way of doing it, and it shows
some issues with label/symbol + offset generation that I need to
go and make happen.  But it does work.

The challenge for strings is that I want to consume more data than
is in t[T_SIZE]. I will need to figure out how to continue iterating
inside the current string block and generate characters rather than
doing it 6 (or now 8!) bytes at a time.  So, stay tuned.